requirement.txt
amqp==5.2.0
billiard==4.2.0
blinker==1.6.3
celery==5.3.6
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
dlib==19.24.2
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask==3.0.0
Flask-Cors==4.0.0
Flask-MySQLdb==2.0.0
greenlet==3.0.0
gunicorn==21.2.0
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
kombu==5.3.4
MarkupSafe==2.1.3
mysqlclient==2.2.0
numpy==1.26.1
packaging==23.2
Pillow==10.1.0
prompt-toolkit==3.0.41
python-dateutil==2.8.2
python-ulid==2.2.0
redis==5.0.1
scikit-learn==1.3.2
scipy==1.11.3
six==1.16.0
SQLAlchemy==2.0.22
threadpoolctl==3.2.0
typing_extensions==4.8.0
tzdata==2023.3
vine==5.1.0
wcwidth==0.2.12
Werkzeug==3.0.1
